Process tags pagination from last to first page
Chaging the order the pagination is processed to avoid problems like
deleting the items in first page and that would make the next page have less items
and in some cases make the next page to disapear.
By processing from the last page to the first one this problem does
not happen.